refactor: deck 클래스로 변경
