+++ src/main/kotlin/blackjack/controller/BlackjackController.kt
@@ -1,7 +1,8 @@
 package blackjack.controller
 
 import blackjack.model.Dealer
-import blackjack.model.Deck.INITIAL_HAND_OUT_CARD_COUNT
+import blackjack.model.Deck
+import blackjack.model.Deck.Companion.INITIAL_HAND_OUT_CARD_COUNT
 import blackjack.model.DrawChoice
 import blackjack.model.GameManager
 import blackjack.model.Player
@@ -13,6 +14,7 @@ class BlackjackController(
     private val outputView: OutputView,
 ) {
     private lateinit var gameManager: GameManager
+
 
     fun play(dealer: Dealer) {
         val players = playerSetting()
