+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -2,7 +2,7 @@ package blackjack.view
 
 import blackjack.model.Card
 import blackjack.model.Dealer
-import blackjack.model.Deck.INITIAL_HAND_OUT_CARD_COUNT
+import blackjack.model.Deck.Companion.INITIAL_HAND_OUT_CARD_COUNT
 import blackjack.model.Player
 import blackjack.model.ResultType
 import blackjack.model.Shape
